make DataLayer its own capability
