+++ packages/teams-js/src/internal/telemetry.ts
@@ -195,7 +195,6 @@ export enum ApiName {
   Menus_SetNavBarMenu = 'menus.setNavBarMenu',
   Menus_SetUpViews = 'menus.setUpViews',
   Menus_ShowActionMenu = 'menus.showActionMenu',
-  MessageChannels_GetCentralDataLayerPort = 'messageChannels.getCentralDataLayerPort',
   MessageChannels_GetTelemetryPort = 'messageChannels.getTelemetryPort',
   Monetization_OpenPurchaseExperience = 'monetization.openPurchaseExperience',
   Navigation_NavigateBack = 'navigation.navigateBack',
@@ -237,6 +236,7 @@ export enum ApiName {
   Pages_Tabs_GetTabInstances = 'pages.tabs.getTabInstances',
   Pages_Tabs_NavigateToTab = 'pages.tabs.navigateToTab',
   People_SelectPeople = 'people.selectPeople',
+  PrivateAPIs_GetCentralDataLayerPort = 'getCentralDataLayerPort',
   PrivateAPIs_OpenFilePreview = 'openFilePreview',
   PrivateAPIs_RegisterCustomHandler = 'registerCustomHandler',
   PrivateAPIs_RegisterUserSettingsChangeHandler = 'registerUserSettingsChangeHandler',
