+++ packages/teams-js/src/public/runtime.ts
@@ -226,6 +226,7 @@ interface IRuntimeV4 extends IBaseRuntime {
     readonly barCode?: {};
     readonly calendar?: {};
     readonly call?: {};
+    readonly centralDataLayer?: {};
     readonly chat?: {};
     readonly clipboard?: {};
     readonly conversations?: {};
